+++ plugkit/src/range.hpp
@@ -1,18 +1,18 @@
 #ifndef PLUGKIT_RANGE_H
 #define PLUGKIT_RANGE_H
 
-#include <stddef.h>
+#include <cstdint>
 
-#define RANGE_NPOS ((size_t)-1)
+#define RANGE_NPOS ((uint32_t)-1)
 
 namespace plugkit {
 
 struct Range {
-  size_t begin;
-  size_t end;
+  uint32_t begin;
+  uint32_t end;
 };
 
-inline Range Range_offset(Range base, size_t offset) {
+inline Range Range_offset(Range base, uint32_t offset) {
   Range range = {base.begin + offset, base.end + offset};
   return range;
 }
